AVRO-915. Large number of warnings in C++ buildsc
git-svn-id: https://svn.apache.org/repos/asf/avro/trunk@1182454 13f79535-47bb-0310-9956-ffa450edef68

+++ lang/c++/api/Stream.hh
@@ -170,7 +170,7 @@ struct StreamReader {
     }
 
     void skipBytes(size_t n) {
-        if (n > (end_ - next_)) {
+        if (n > static_cast<size_t>(end_ - next_)) {
             n -= end_ - next_;
             next_ = end_;
             in_->skip(n);
